requirements.txt
customtkinter
endesive
cryptography
Pillow
